chore: ponytail-audit cleanup (remove yagni interfaces and hoplite dependency)
