+++ src/main/kotlin/net/raquezha/nuecagram/Config.kt
@@ -1,5 +1,8 @@
 package net.raquezha.nuecagram
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class Config(
     val name: String,
     val env: String,
