--- src/main/kotlin/net/raquezha/nuecagram/di/SystemEnv.kt
@@ -1,7 +1,0 @@
-package net.raquezha.nuecagram.di
-
-interface SystemEnv {
-    fun getBotApi(): String
-
-    fun getSecretToken(): String
-}
