--- src/main/kotlin/net/raquezha/nuecagram/telegram/TokenProvider.kt
@@ -1,7 +1,0 @@
-package net.raquezha.nuecagram.telegram
-
-interface TokenProvider {
-    fun getBotToken(): String
-
-    fun getSecretToken(): String
-}
