--- src/main/kotlin/net/raquezha/nuecagram/webhook/RandomMessageProviderImpl.kt
@@ -1,239 +1,0 @@
-package net.raquezha.nuecagram.webhook
-
-/**
- * Implementation of [RandomMessageProvider] that provides humorous random messages
- * for pipeline completion notifications.
- *
- * Each category contains 50 messages to provide variety in notifications.
- */
-class RandomMessageProviderImpl : RandomMessageProvider {
-    companion object {
-        private val SUCCESS_MESSAGES =
-            listOf(
-                "Stop sipping that coffee, pipeline passed!",
-                "Pipeline passed! You're officially a 10x developer today.",
-                "Pipeline passed! Even your code is surprised.",
-                "Success! The CI gods have smiled upon you.",
-                "Pipeline passed! Quick, deploy before someone breaks it!",
-                "It worked?! I mean... of course it worked!",
-                "All green! Your code is chef's kiss today.",
-                "Pipeline passed! Time to mass produce those features!",
-                "Green across the board! You're on fire today!",
-                "Pipeline passed! Your future self thanks you.",
-                "Success! The code review gods are pleased.",
-                "Pipeline passed! Screenshot this moment.",
-                "All tests passed! Is this real life?",
-                "Pipeline passed! You've earned that coffee break.",
-                "Success! The build fairy visited your code.",
-                "Pipeline passed! No bugs were harmed in this commit.",
-                "Green light! Your code made it through the gauntlet.",
-                "Pipeline passed! Time to celebrate responsibly.",
-                "Success! Your keyboard deserves a medal.",
-                "Pipeline passed! The CI gremlins are sleeping today.",
-                "Holy semicolons! The pipeline actually passed!",
-                "Pipeline passed! Someone notify the historians!",
-                "Success! Your code just graduated with honors.",
-                "Pipeline passed! The hamsters powering the server are proud.",
-                "Witchcraft detected! Pipeline passed on first try!",
-                "Pipeline passed! Quick, buy a lottery ticket!",
-                "Success! The code pixies blessed your commit.",
-                "Pipeline passed! Your ancestors are smiling.",
-                "Achievement unlocked: Pipeline Whisperer!",
-                "Pipeline passed! The prophecy has been fulfilled!",
-                // 10 light humor messages
-                "Pipeline passed! Treat yourself to a snack.",
-                "Success! Your code is having a good hair day.",
-                "Pipeline passed! The weekend just got better.",
-                "Nice work! The pipeline gave you a virtual high-five.",
-                "Pipeline passed! Your plant is proud of you too.",
-                "Success! Time for a tiny desk celebration.",
-                "Pipeline passed! Even Monday feels okay now.",
-                "Well done! The pipeline is doing a happy dance.",
-                "Pipeline passed! Your rubber duck approves.",
-                "Success! Go tell your pet about this achievement.",
-                // 10 Filipino dev meme messages
-                "Pipeline passed! Isa ka nang alamat, par!",
-                "Success! Sana all ganito ka-smooth ang build!",
-                "Pipeline passed! Tapos na, uwian na!",
-                "Pipeline passed! Pa-Jollibee ka naman diyan!",
-                "Success! Grabe, first try pa! Baka naman!",
-                "Pipeline passed! Tuloy ang laro, dev!",
-                "Galing mo, idol! Pipeline passed na passed!",
-                "Pipeline passed! Pang-senior dev ka na niyan!",
-            )
-
-        private val FAILED_MESSAGES =
-            listOf(
-                "The pipeline has passed... away. RIP.",
-                "Pipeline failed! Time to check Stack Overflow.",
-                "Pipeline failed! But hey, at least you're consistent.",
-                "Build machine said: 'Nah, I don't think so.'",
-                "Pipeline failed! Have you tried turning it off and on again?",
-                "Failed! The code gods demand a sacrifice (your lunch break).",
-                "Pipeline failed! Time for git blame!",
-                "Oops! Your code took the day off. Pipeline failed!",
-                "Pipeline failed! May the force rebuild with you.",
-                "Failed! Time to enter debug mode.",
-                "Pipeline failed! The rubber duck demands answers.",
-                "Build failed! Your code needs a hug.",
-                "Pipeline failed! Console.log to the rescue!",
-                "Failed! Time to add more coffee to the code.",
-                "Pipeline failed! The tests have spoken.",
-                "Build broke! Time to channel your inner detective.",
-                "Pipeline failed! Your code is playing hard to get.",
-                "Failed! The merge conflicts send their regards.",
-                "Pipeline failed! Even the linter is disappointed.",
-                "Build failed! Time to question all life choices.",
-                "Pipeline failed! Your code has chosen chaos.",
-                "Build exploded! Somewhere, a semicolon is laughing.",
-                "Pipeline failed! The server room felt a disturbance.",
-                "Failed! Your code went full gremlin mode.",
-                "Pipeline failed! Even the interns are shaking their heads.",
-                "Build failed! The error messages have formed a union.",
-                "Pipeline failed! Your code is crying in binary.",
-                "Failed! The null pointer sends its regards.",
-                "Pipeline failed! Looks like the code had too much coffee.",
-                "Build failed! Time to blame cosmic radiation.",
-                // 10 light humor messages
-                "Pipeline failed! Nothing a good snack can't fix.",
-                "Oops! The pipeline tripped on a shoelace.",
-                "Pipeline failed! Take a deep breath, you've got this.",
-                "Failed! Even the best have off days.",
-                "Pipeline failed! The code just needs a little nap.",
-                "Build hiccup! Rome wasn't built in one pipeline either.",
-                "Pipeline failed! Time for a stretch break.",
-                "Failed! Your code is just being dramatic today.",
-                "Pipeline failed! Plot twist in your coding journey.",
-                "Oops! The pipeline stubbed its toe. Try again!",
-                // 10 Filipino dev meme messages
-                "Pipeline failed! Bahala na si Batman!",
-                "Pipeline failed! 'Di bale, marami pang bukas!",
-                "Ay jusko! Sino nag-push nito?!",
-                "Failed! Kape muna tayo, par. Mahaba-habang gabi 'to.",
-                "Pipeline failed! Bawi na lang next sprint!",
-                "Failed! Sana ol may working code!",
-                "Pipeline failed! Eto na naman tayo, debug ulit!",
-                "Pipeline failed! Tara na sa Stack Overflow!",
-            )
-
-        private val CANCELED_MESSAGES =
-            listOf(
-                "Pipeline canceled! Someone got cold feet.",
-                "Pipeline canceled! Commitment issues detected.",
-                "Canceled! The pipeline ghosted you.",
-                "Pipeline canceled! It's not you, it's the code.",
-                "Abort mission! Pipeline canceled!",
-                "Pipeline canceled! Strategic retreat initiated.",
-                "Canceled! The pipeline needed a mental health day.",
-                "Pipeline canceled! Plot twist avoided.",
-                "Canceled! Someone pulled the emergency brake.",
-                "Pipeline canceled! Tactical withdrawal complete.",
-                "Pipeline canceled! The undo button was pressed.",
-                "Canceled! The pipeline got stage fright.",
-                "Pipeline canceled! Back to the drawing board.",
-                "Canceled! The pipeline called in sick.",
-                "Pipeline canceled! Escape pod deployed.",
-                "Canceled! The pipeline took an early exit.",
-                "Pipeline canceled! Reality check requested.",
-                "Canceled! The pipeline needed more time to think.",
-                "Pipeline canceled! Ctrl+Z accepted.",
-                "Canceled! The pipeline chose self-preservation.",
-                "Pipeline canceled! Someone panicked and hit the big red button.",
-                "Canceled! The pipeline saw its future and noped out.",
-                "Pipeline canceled! It's having an existential crisis.",
-                "Canceled! The pipeline went to get milk and never came back.",
-                "Pipeline canceled! Witness protection program activated.",
-                "Canceled! The pipeline joined the witness protection program.",
-                "Pipeline canceled! It saw the code and ran away screaming.",
-                "Canceled! Pipeline went out for a smoke break... forever.",
-                "Pipeline canceled! It's not ghosting, it's strategic absence.",
-                "Canceled! The pipeline filed for early retirement.",
-                // 10 light humor messages
-                "Pipeline canceled! Change of plans, no worries.",
-                "Canceled! The pipeline is taking a breather.",
-                "Pipeline canceled! Sometimes pause is progress.",
-                "Canceled! Good call, better safe than sorry.",
-                "Pipeline canceled! The universe said 'not yet.'",
-                "Canceled! Smart move, reassess and retry.",
-                "Pipeline canceled! Taking the scenic route instead.",
-                "Canceled! The pipeline is rethinking life choices.",
-                "Pipeline canceled! Intermission time.",
-                "Canceled! The pipeline requested a timeout.",
-                // 10 Filipino dev meme messages
-                "Pipeline canceled! Petiks muna, par!",
-                "Canceled! Nagbago ang isip ni boss!",
-                "Pipeline canceled! Ayaw na pala ni client!",
-                "Canceled! Teka lang, may revision daw!",
-                "Pipeline canceled! Change request na naman!",
-                "Canceled! Wag muna, meeting daw ulit!",
-                "Pipeline canceled! Sabi ni PM, hold muna!",
-                "Canceled! Last minute changes strikes again!",
-                "Pipeline canceled! Nag-leave yung nag-approve!",
-                "Canceled! Balik drawing board, sabi ng stakeholder!",
-            )
-
-        private val SKIPPED_MESSAGES =
-            listOf(
-                "Pipeline skipped! It said 'not today.'",
-                "Skipped! The pipeline is feeling lazy.",
-                "Pipeline skipped! Maybe next time?",
-                "Skipped! Your pipeline is on vacation mode.",
-                "Pipeline said 'skip' like it's a YouTube ad.",
-                "Skipped! The pipeline is practicing mindfulness.",
-                "Pipeline skipped! It's taking a rain check.",
-                "Skipped! The pipeline is conserving energy.",
-                "Pipeline skipped! Speed run mode activated.",
-                "Skipped! The pipeline knows a shortcut.",
-                "Pipeline skipped! Express lane taken.",
-                "Skipped! The pipeline is being efficient.",
-                "Pipeline skipped! Fast forward enabled.",
-                "Skipped! The pipeline trusts your code.",
-                "Pipeline skipped! Nothing to see here.",
-                "Skipped! The pipeline is on autopilot.",
-                "Pipeline skipped! Turbo mode engaged.",
-                "Skipped! The pipeline took the scenic route... around.",
-                "Pipeline skipped! Optimization level: maximum.",
-                "Skipped! The pipeline is ahead of schedule.",
-                "Pipeline skipped! It's too cool for this commit.",
-                "Skipped! The pipeline has better things to do.",
-                "Pipeline skipped! It's having a 'me' day.",
-                "Skipped! The pipeline is binge-watching Netflix instead.",
-                "Pipeline skipped! It left the chat.",
-                "Skipped! The pipeline is playing hooky.",
-                "Pipeline skipped! It's vibing in another dimension.",
-                "Skipped! The pipeline said 'that's above my pay grade.'",
-                "Pipeline skipped! It's in its avoidance era.",
-                "Skipped! The pipeline took a personal day. No further questions.",
-                // 10 light humor messages
-                "Pipeline skipped! Saving energy for later.",
-                "Skipped! The pipeline is pacing itself.",
-                "Pipeline skipped! Working smarter, not harder.",
-                "Skipped! The pipeline believes in you already.",
-                "Pipeline skipped! It read the code and said 'looks good!'",
-                "Skipped! No need to run, confidence is high.",
-                "Pipeline skipped! Taking the express checkout.",
-                "Skipped! The pipeline is delegating today.",
-                "Pipeline skipped! Fast pass activated.",
-                "Skipped! The pipeline gave you a free pass.",
-                // 10 Filipino dev meme messages
-                "Pipeline skipped! Bahala na bukas!",
-                "Skipped! Wala namang nagtatanong, skip na!",
-                "Pipeline skipped! Chill lang muna, par!",
-                "Skipped! Para sa mental health!",
-                "Pipeline skipped! Friday na kasi eh!",
-                "Skipped! 'Di naman priority 'yan sabi ni boss!",
-                "Pipeline skipped! Next sprint na lang daw!",
-                "Skipped! Out of scope na daw!",
-                "Pipeline skipped! Backlog na lang muna!",
-                "Skipped! Pwede na 'yan, ship na!",
-            )
-    }
-
-    override fun getSuccessMessage(): String = "✅ ${SUCCESS_MESSAGES.random()}"
-
-    override fun getFailedMessage(): String = "❌ ${FAILED_MESSAGES.random()}"
-
-    override fun getCanceledMessage(): String = "⛔ ${CANCELED_MESSAGES.random()}"
-
-    override fun getSkippedMessage(): String = "⏭️ ${SKIPPED_MESSAGES.random()}"
-}
